% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{execute}
\alias{execute}
\title{Execute HivDescriptive Study}
\usage{
execute(connectionDetails, connp, studyp, tablePrefix = "", outputFolder,
  createCohorts = TRUE, cohortTable, runDiagnostics = TRUE,
  packageResults = FALSE)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{tablePrefix}{The prefix for the study tables, should be same value used in init()}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}
}
\description{
Execute HivDescriptive Study
}
\details{
This function executes the HivDescriptive Study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchema = "cdm_data",
        targetDatabaseSchema = "results",
        oracleTempSchema = NULL,
        tablePrefix = "endo_",
        outputFolder = "c:/temp/study_results")
}

}
